data_lake/requirements.txt
tweepy>=4.12.1
python-dotenv>=1.0.0
textblob>=0.17.1
pandas>=1.3.0
matplotlib>=3.4.0
wordcloud>=1.8.0
nltk>=3.6.0
numpy>=1.21.0
Pillow>=10.0.0
PyMuPDF>=1.21.0
